+++ lib/subghz/protocols/faac_slh.c
@@ -111,7 +111,7 @@ void subghz_protocol_decoder_faac_slh_feed(void* context, bool level, uint32_t d
         break;
     case FaacSLHDecoderStepSaveDuration:
         if(level) {
-            if(duration >= (subghz_protocol_faac_slh_const.te_short * 3 +
+            if(duration >= ((uint32_t)subghz_protocol_faac_slh_const.te_short * 3 +
                             subghz_protocol_faac_slh_const.te_delta)) {
                 instance->decoder.parser_step = FaacSLHDecoderStepFoundPreambula;
                 if(instance->decoder.decode_count_bit >=
