+++ lib/subghz/protocols/scher_khan.c
@@ -150,7 +150,7 @@ void subghz_protocol_decoder_scher_khan_feed(void* context, bool level, uint32_t
         break;
     case ScherKhanDecoderStepSaveDuration:
         if(level) {
-            if(duration >= (subghz_protocol_scher_khan_const.te_long +
+            if(duration >= (uint32_t)(subghz_protocol_scher_khan_const.te_long +
                             subghz_protocol_scher_khan_const.te_delta * 2)) {
                 //Found stop bit
                 instance->decoder.parser_step = ScherKhanDecoderStepReset;
